+++ src/mvm/heap.h
@@ -1,5 +1,5 @@
 /* Niels Widger
- * Time-stamp: <03 Apr 2011 at 15:03:53 by nwidger on macros.local>
+ * Time-stamp: <03 Apr 2011 at 16:08:32 by nwidger on macros.local>
  */
 
 #ifndef _MVM_HEAP_H
@@ -41,7 +41,7 @@ void heap_clear(struct heap *h);
  * @return the current size of the heap in bytes
  */ 
 
-int heap_get_size(struct heap *h);
+uint64_t heap_get_size(struct heap *h);
 
 /** returns the current free space in the given heap in bytes.
  *
@@ -50,7 +50,7 @@ int heap_get_size(struct heap *h);
  * @return the current free space in the heap in bytes
  */
 
-int heap_get_free(struct heap *h);
+uint64_t heap_get_free(struct heap *h);
 
 /** allocates b bytes of memory from the given heap and returns a
  * pointer to the allocated space.
